android: [TableGen] Include basic generators in the non-min tblgen
Generators are registered as global objects, so their symbols need to be
included even if unreferenced. Android.bp has whole_static_libs for that goal.

Basic generators are available in llvm-min-tblgen, but we duplicate them in the
non-min variant to avoid commits like 20601a7673b4570caf4ab0028c1607c8b6240d74.

The CMake build does this.